requirements.txt
absl-py==1.0.0
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.10
dill==0.3.5.1
Django==4.0.1
django-json-widget==1.1.1
djangorestframework==3.13.1
future==0.18.2
gunicorn==20.1.0
idna==3.3
loguru==0.6.0
lxml==4.8.0
multiprocess==0.70.12.2
numpy==1.22.1
ortools==9.2.9972
pathos==0.2.8
pox==0.3.1
ppft==1.7.6.5
protobuf==3.19.3
psycopg2-binary==2.9.3
PyExecJS==1.5.1
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tabulate==0.8.9
translators==5.2.1
urllib3==1.26.8
